+++ lib/deprecation_toolkit/configuration.rb
@@ -12,5 +12,35 @@ module DeprecationToolkit
     config_accessor(:deprecation_path) { "test/deprecations" }
     config_accessor(:test_runner) { :minitest }
     config_accessor(:warnings_treated_as_deprecation) { [] }
+
+    config_accessor(:message_normalizers) do
+      normalizers = []
+
+      normalizers << PathPrefixNormalizer.new(Rails.root) if defined?(Rails)
+      normalizers << PathPrefixNormalizer.new(Bundler.root) if defined?(Bundler)
+
+      if defined?(Gem)
+        Gem.loaded_specs.each_value do |spec|
+          normalizers << PathPrefixNormalizer.new(spec.bin_dir, replacement: "<GEM_BIN_DIR:#{spec.name}>")
+          normalizers << PathPrefixNormalizer.new(spec.extension_dir, replacement: "<GEM_EXTENSION_DIR:#{spec.name}>")
+          normalizers << PathPrefixNormalizer.new(spec.gem_dir, replacement: "<GEM_DIR:#{spec.name}>")
+        end
+        normalizers << PathPrefixNormalizer.new(*Gem.path, replacement: "<GEM_PATH>")
+      end
+
+      begin
+        require "rbconfig"
+        normalizers << PathPrefixNormalizer.new(
+          *RbConfig::CONFIG.values_at("prefix", "sitelibdir", "rubylibdir"),
+          replacement: "<RUBY_INTERNALS>",
+        )
+      rescue LoadError
+        # skip normalizing ruby internal paths
+      end
+
+      normalizers << PathPrefixNormalizer.new(Dir.pwd)
+
+      normalizers
+    end
   end
 end